yuju !

+++ igr/practica3/pv2f.cpp
@@ -16,39 +16,39 @@ PV2f::PV2f(const PV2f &pv2f, Tipo tipo)
 {
 }
 
-PV2f::PV2f(GLfloat x, GLfloat y, Tipo tipo)
+PV2f::PV2f(GLdouble x, GLdouble y, Tipo tipo)
     : m_x(x)
     , m_y(y)
     , m_tipo(tipo)
 {
 }
 
-GLfloat PV2f::getX() const
+GLdouble PV2f::getX() const
 {
     return m_x;
 }
 
-GLfloat PV2f::getY() const
+GLdouble PV2f::getY() const
 {
     return m_y;
 }
 
-GLfloat &PV2f::rX()
+GLdouble &PV2f::rX()
 {
     return m_x;
 }
 
-GLfloat &PV2f::rY()
+GLdouble &PV2f::rY()
 {
     return m_y;
 }
 
-void PV2f::setX(GLfloat x)
+void PV2f::setX(GLdouble x)
 {
     m_x = x;
 }
 
-void PV2f::setY(GLfloat y)
+void PV2f::setY(GLdouble y)
 {
     m_y = y;
 }
@@ -84,14 +84,14 @@ void PV2f::normalizar()
     m_y *= c;
 }
 
-void PV2f::rotar(const PV2f &centro, GLfloat angulo)
+void PV2f::rotar(const PV2f &centro, GLdouble angulo)
 {
-    const GLfloat cx = centro.getX();
-    const GLfloat cy = centro.getY();
-    const GLfloat dx = (1.0 - cos(angulo)) * cx + sin(angulo) * cy;
-    const GLfloat dy = -sin(angulo) * cx + (1.0 - cos(angulo)) * cy;
-    const GLfloat x = m_x * cos(angulo) - m_y * sin(angulo) + dx;
-    const GLfloat y = m_x * sin(angulo) + m_y * cos(angulo) + dy;
+    const GLdouble cx = centro.getX();
+    const GLdouble cy = centro.getY();
+    const GLdouble dx = (1.0 - cos(angulo)) * cx + sin(angulo) * cy;
+    const GLdouble dy = -sin(angulo) * cx + (1.0 - cos(angulo)) * cy;
+    const GLdouble x = m_x * cos(angulo) - m_y * sin(angulo) + dx;
+    const GLdouble y = m_x * sin(angulo) + m_y * cos(angulo) + dy;
     m_x = x;
     m_y = y;
 }
